Update README.md
Hello everyone! This is the coding project that I, and an expansive team of experience developers (5 15 year olds) will be attempting to solve this summer. This is a software built for all users to manage and organise all kinds of tasks and events in a user friendly and simple GUI. Our aim is to help our users maximise productivity with zero costs!
This is our Team:
*Tanay Soni*: Project Manager, Main Coder, Reviewer and Editor
*Ritwik Kumar*: Project Manager, Main Coder and Debugger
*Arpit Verma*: Main Coder, Reviewer and Editor
*Vedant Upadhyay*: Code Tester and Head of the Gooning Department
*Shubh Gordhan*: Lead App and Interface Designer